Fix default pattern
