+++ .changeset/selfish-cougars-sniff.md
@@ -1,0 +1,5 @@
+---
+"@neo4j/cypher-builder": minor
+---
+
+Support for complex patterns
