+++ packages/cypher-builder/tests/patterns.test.ts
@@ -46,7 +46,7 @@ describe("Patterns", () => {
 
             const query = new TestClause(a.related(rel).to(b));
             const queryResult = query.build();
-            expect(queryResult.cypher).toMatchInlineSnapshot(`"(this0)-[this1:\`ACTED_IN\`]->(this2)"`);
+            expect(queryResult.cypher).toMatchInlineSnapshot(`"(this0)-[this1:ACTED_IN]->(this2)"`);
 
             expect(queryResult.params).toMatchInlineSnapshot(`Object {}`);
         });
@@ -70,7 +70,7 @@ describe("Patterns", () => {
             const query = new TestClause(a.related(rel).to(b));
             const queryResult = query.build();
             expect(queryResult.cypher).toMatchInlineSnapshot(
-                `"(this0:\`Person\`:\`Actor\` { name: $param0, surname: $param1 })-[this1:\`ACTED_IN\` { roles: $param2 }]->(this2)"`
+                `"(this0:\`Person\`:\`Actor\` { name: $param0, surname: $param1 })-[this1:ACTED_IN { roles: $param2 }]->(this2)"`
             );
 
             expect(queryResult.params).toMatchInlineSnapshot(`
@@ -99,7 +99,7 @@ describe("Patterns", () => {
             const query = new TestClause(a.related(rel1).to(b).related(rel2).to(c));
             const queryResult = query.build();
             expect(queryResult.cypher).toMatchInlineSnapshot(
-                `"(this0)-[this1:\`ACTED_IN\`]->(this2)-[this3:\`ACTED_IN\`]->(this4:\`TestLabel\`)"`
+                `"(this0)-[this1:ACTED_IN]->(this2)-[this3:ACTED_IN]->(this4:\`TestLabel\`)"`
             );
 
             expect(queryResult.params).toMatchInlineSnapshot(`Object {}`);
